pyproject.toml
[tool.poetry]
name = "tgquizbot"
version = "0.1.0"
description = ""
authors = ["Myakot <cstepan@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
pyTelegramBotAPI = "^4.21.0"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
